Mod 'edit_note()' and 'delete_note()'

+++ task3_note_app.py
@@ -1,5 +1,5 @@
 import os
-from colorama import init, Fore, Style
+from colorama import init, Fore, Style  # добавляем колораму, дабы текст в консоли был читабельнее
 init(autoreset=True)
 
 
@@ -16,7 +16,7 @@ def create_note(note_name):
 
 
 def read_note(note_name):
-    if os.path.isfile(f"{note_name}.txt"):
+    if os.path.isfile(f"{note_name}.txt"):  # проверяем, существует ли файл
         with open(f"{note_name}.txt", "r", encoding="utf-8") as file:
             note_text = file.read()
         print(Style.BRIGHT + Fore.GREEN + f'Текст заметки: {note_text}')
@@ -25,31 +25,36 @@ def read_note(note_name):
 
 
 def edit_note(note_name):
-    if os.path.isfile(f"{note_name}.txt"):
-
+    if os.path.isfile(f"{note_name}.txt"):  # проверяем, существует ли файл
         # Читаем файл и выводим на экран
-        with open(f"{note_name}.txt", "r", encoding="utf-8") as file:
-            note_text = file.read()
-        print(Style.BRIGHT + Fore.BLUE + f'Старый текст заметки: {note_text}')
-
-        # Перезаписываем файл и тоже выводим на экран
-        with open(f"{note_name}.txt", "w+", encoding="utf-8") as file:
+        print(Style.BRIGHT + Fore.YELLOW + 'Старый', end = ' ')
+        read_note(note_name)
+        # Перезаписываем файл
+        with open(f"{note_name}.txt", "w", encoding="utf-8") as file:
             note_text = input('Введите новый текст: ')
             file.write(note_text)
-            #print(Style.BRIGHT + Fore.GREEN + f'Новый текст заметки: {file.read()}\n')
-        print(Style.BRIGHT + Fore.GREEN + 'Новый', end = ' ')
+        # Выводим на экран перезаписанный файл
+        print(Style.BRIGHT + Fore.YELLOW + 'Новый', end = ' ')
         read_note(note_name)
         print(Style.BRIGHT + Fore.CYAN + 'Текст заметки успешно перезаписан.')
     else: 
         print(Fore.RED + 'Заметка не найдена.')
 
 
-def delete_note(note_name):
-    if os.path.isfile(f"{note_name}.txt"):
-        os.remove(f"{note_name}.txt")
-        print(Style.BRIGHT + Fore.RED + 'Файл удалён!')
+def delete_note(note_name):    
+    if os.path.isfile(f"{note_name}.txt"):  # проверяем, существует ли файл
+        thirst_for_deletion = input(
+            "\nВы действительно хотите удалить файл?\n"
+            "Если да, введите латинскую 'y', если нет - то произвольные символы: "
+            )
+        if thirst_for_deletion == 'y':
+            os.remove(f"{note_name}.txt")
+            print(Style.BRIGHT + Fore.RED + 'Файл удалён!')
+        else:
+            print(Style.BRIGHT + Fore.CYAN + 'Удаление отменено. Возврат в Основное Меню.')    
     else: 
         print(Fore.RED + 'Заметка не найдена.')
+    
 
 
 def main():
@@ -62,16 +67,19 @@ def main():
             '4': 'delete_note',
         }
         print('')   # добавляем одиночный отступ
+        
+        # Вывод меню
         for k in sorted(menu.keys()):
             print(f'{k} : {menu.get(k)}')
         
-        key = input('\nВведите ключ операции. Если хотите выйти, введите любые другие символы: ',)
+        key = input('\nВведите ключ операции. Если хотите выйти, введите произвольные символы: ')
         if key == '1' or key == '2' or key == '3' or key == '4':
             print(Style.BRIGHT + Fore.CYAN + f'Выбрана операция: {menu[key]}\n')
             note_name = input("Введите название заметки: ")
         else:
             print(Style.BRIGHT + Fore.CYAN + 'Работа программы завершена.\n')
 
+        # Вызов соответствующей операции или завершение программы
         if key == '1':
             create_note(note_name)
         elif key == '2':
